third_party/WebKit/LayoutTests/nfc/mock-nfc.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/services/device/public/mojom/nfc.mojom.js"></script>
<script src="resources/nfc-helpers.js"></script>
<script>

'use strict';

nfc_test(() => {
  assert_true(mockNFC instanceof Object);
}, 'NFC Mojo bindings and mock interfaces are available to tests.')
</script>
